lab02 refactor remove home module.
