+++ json-server/db.json
@@ -33,7 +33,7 @@
       "title": "Introduction to Javascript",
       "url": "https://www.sample-videos.com/video123/mp4/720/big_buck_bunny_720p_1mb.mp4",
       "duration": "0:13",
-      "approved": false
+      "approved": true
     }
   ]
 }
