+++ lab02_angular8/smart-player/src/app/app-routing.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomeComponent } from './home/home.component';
-import { CourseComponent } from './course/course.component';
+import { CourseComponent } from './course';
+import { VideoPlayerComponent } from './video-player/video-player.component';
 
 const routes: Routes = [
   { path: '', redirectTo: 'course', pathMatch: 'full' },
-  { path: 'watch', component: HomeComponent, pathMatch: 'full' },
+  { path: 'watch', component: VideoPlayerComponent, pathMatch: 'full' },
   { path: 'course', component: CourseComponent, pathMatch: 'full' },
   { path: '**', component: CourseComponent },
 ];
