--- lab02_angular8/smart-player/src/app/home/home.component.html
@@ -1,1 +1,0 @@
-<app-video-player></app-video-player>
