--- lab02_angular8/smart-player/src/app/home/video-player/index.ts
@@ -1,4 +1,0 @@
-export * from './video-player.component';
-export * from './playlist';
-export * from './player';
-export * from './controls';
