+++ lab02_angular8/smart-player/src/app/video-player/components/index.ts
@@ -1,0 +1,3 @@
+export * from './player';
+export * from './controls';
+export * from './playlist';
