+++ lab02_angular8/smart-player/src/app/video-player/components/player/player.component.html
