+++ lab02_angular8/smart-player/src/app/video-player/index.ts
@@ -1,0 +1,1 @@
+export * from './video-player.module';
