+++ lab02_angular8/smart-player/src/app/video-player/video-player.component.css
